NFC: move the remaining protocols' scenes into their plugins
Same treatment as Ultralight, for the six protocols that were left: MIFARE
Classic (dictionary attack, detect reader, both mfkey screens, show keys,
update initial), MIFARE Plus (dictionary attack, show keys, more info, ISO4
info, version, update initial), SLIX (unlock menu, key input, unlock, unlock
success), FeliCa (more info, system), DESFire (more info, app) and EMV
(transactions). Twenty-three scenes, no logic changed, thunks left in the app so
every scene id and call site still resolves.

Two thirds of the saving is indirect. Five render objects -
mf_desfire_render, felica_render, mf_plus_render, iso14443_4a_render and
emv_render, 6,839 B - were only in the app image because resident scenes called
them, and they were already compiled into the .fal as well. With the scenes gone
they garbage-collect out, and the duplication goes with them. mfkey32_logger and
slix_unlock follow their only callers into nfc_mf_classic.fal and nfc_slix.fal,
and EMV's 64-bit divide takes libgcc's _udivmoddi4 with it.

Scenes reachable before any card is read now name the protocol they need, via
nfc_protocol_support_load(): Extra Actions for both unlock menus, the start
menu's Detect Reader, and the two save paths that jump to the mfkey screens.
Without this the Ultralight unlock menu would have hit the "Plugin Not Loaded"
screen - a regression the Ultralight commit introduced and this one fixes.

Mechanical adjustments, all forced by the scenes sharing one translation unit
per protocol: handlers that only did UNUSED(context) now take NfcApp*, and the
`enum SubmenuIndex` that FeliCa and DESFire each declared twice with different
members is scoped per scene.

The app manifest excludes *_extra_scenes.c rather than naming each file, and
gives nfc_mf_classic and nfc_slix the helpers their new scenes need.

Resident image: 57,483 -> 32,927 B (-24,556). Full arc: 77,226 -> 32,927
(-44,299, -57.4%). Scene count is unchanged at 67 - the thunks stay.

Plugins grow, and are only resident while that card is loaded: Ultralight
9,610 -> 13,750, Classic 4,373 -> 11,778, Plus 3,083 -> 10,509, DESFire
3,369 -> 5,653, SLIX 3,376 -> 4,852, FeliCa 1,685 -> 4,362, EMV 1,096 -> 2,672.
Worst case, an Ultralight card loaded, is 46,677 B against 86,836 before.
